+++ CHANGELOG.md
@@ -14,6 +14,33 @@ beta releases are not included in this history.
 
 [//]: # (begin_release_notes)
 
+"1.109.1" (2025-08-01)
+======================
+
+Bugfixes
+--------
+
+- Dependency upgrade: v4-4.25.7 (#20250801105313, #20250801105328, #20250801105343, #20250801105356, #20250801105410, #20250801105424, #20250801105438, #20250801105448, #20250801105504, #20250801105514)
+- :bug: `filesystem` IsFile should return true for special unix files not just regular files (#20250801151058)
+
+
+"1.109.0" (2025-08-01)
+======================
+
+Features
+--------
+
+- :sparkles: `commonerrors` Add IgnoreCorrespondTo function (#20250801105019)
+
+
+Bugfixes
+--------
+
+- :bug: `[platform]` make sure command with privileges are run when passwords are not needed (#20250731182237)
+- :bug: `[filesystem]` make sure to unlink files when removal is called so that sockets are correctly closed (#20250731182404)
+- :bug: `filsystem` Unlink files before attempting to remove them in Remove (#20250801093542)
+
+
 "1.108.2" (2025-07-30)
 ======================